[UNI-353] refactor : 패키지 컨밴션에 맞도록 이동
